Remove duplication checks
